pyproject.toml
[project]
name = "listener"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]

[tool.poetry]
name = "your_project_name"
version = "0.1.0"
description = ""
packages = [{ include = "listener" }]

[tool.poetry.group.dev.dependencies]
discord = "^2.3.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
